LambdaExpression ⇢ CompoundExpression
