+++ src/style-spec/function/expression.js
@@ -3,10 +3,6 @@
 const { typename } = require('./types');
 
 import type { Type } from './types';
-import type { ExpressionName } from './expression_name';
-
-export type NArgs = { kind: 'nargs', types: Array<Type>, N: number };
-export type Signature = Array<Type | NArgs>;
 
 export type CompileError = {|
     error: string,
@@ -16,6 +12,8 @@ export type CompileError = {|
 export interface Expression {
     key: string;
     type: Type;
+
+    static parse(args: Array<mixed>, context: ParsingContext): Expression;
 
     compile(): string | Array<CompileError>;
     serialize(): any;
@@ -61,7 +59,7 @@ class ParsingContext {
     key: string;
     path: Array<number>;
     ancestors: Array<string>;
-    definitions: {[string]: Class<LambdaExpression>};
+    definitions: {[string]: Class<Expression>};
     scope: Scope;
     constructor(definitions: *, path: Array<number> = [], ancestors: * = [], scope: Scope = new Scope()) {
         this.definitions = definitions;
@@ -76,75 +74,6 @@ class ParsingContext {
         const ancestors = expressionName ? this.ancestors.concat(expressionName) : this.ancestors;
         const scope = bindings ? this.scope.concat(bindings) : this.scope;
         return new ParsingContext(this.definitions, path, ancestors, scope);
-    }
-}
-
-class LambdaExpression implements Expression {
-    key: string;
-    type: Type;
-    args: Array<Expression>;
-
-    constructor(key: *, type: Type, args: Array<Expression>) {
-        this.key = key;
-        this.type = type;
-        this.args = args;
-    }
-
-    applyType(type: Type, args: Array<Expression>): Expression {
-        return new this.constructor(this.key, type, args);
-    }
-
-    compile(): string | Array<CompileError> {
-        const errors: Array<CompileError> = [];
-        const compiledArgs: Array<string> = [];
-
-        const args = this.args;
-        for (let i = 0; i < args.length; i++) {
-            const arg = args[i];
-            const compiledArg = arg.compile();
-            if (Array.isArray(compiledArg)) {
-                errors.push.apply(errors, compiledArg);
-            } else {
-                compiledArgs.push(`(${compiledArg})`);
-            }
-        }
-
-        if (errors.length > 0) {
-            return errors;
-        }
-
-        return this.compileFromArgs(compiledArgs);
-    }
-
-    compileFromArgs(_: Array<string>): string | Array<CompileError> {
-        throw new Error('Unimplemented');
-    }
-
-    serialize() {
-        const name = this.constructor.opName();
-        const args = this.args.map(e => e.serialize());
-        return [ name ].concat(args);
-    }
-
-    visit(fn: (Expression) => void) {
-        fn(this);
-        this.args.forEach(a => a.visit(fn));
-    }
-
-    // implemented by subclasses
-    static opName(): ExpressionName { throw new Error('Unimplemented'); }
-    static type(): Type { throw new Error('Unimplemented'); }
-    static signatures(): Array<Signature> { throw new Error('Unimplemented'); }
-
-    // default parse; overridden by some subclasses
-    static parse(args: Array<mixed>, context: ParsingContext): LambdaExpression {
-        const op = this.opName();
-        const parsedArgs: Array<Expression> = [];
-        for (const arg of args) {
-            parsedArgs.push(parseExpression(arg, context.concat(1 + parsedArgs.length, op)));
-        }
-
-        return new this(context.key, this.type(), parsedArgs);
     }
 }
 
@@ -202,20 +131,10 @@ function parseExpression(expr: mixed, context: ParsingContext) : Expression {
     }
 }
 
-function nargs(N: number, ...types: Array<Type>) : NArgs {
-    return {
-        kind: 'nargs',
-        types,
-        N
-    };
-}
-
 module.exports = {
     Scope,
     ParsingContext,
     ParsingError,
     parseExpression,
-    LambdaExpression,
-    Reference,
-    nargs
+    Reference
 };
